coding agent v1/coder-buddy/requirements.txt
requests>=2.28.0
pydantic>=1.10.0
python-dotenv>=0.19.0
langchain>=0.1.0
langgraph>=0.1.0
langchain-ollama>=0.1.0
langchain-openai>=0.1.0